extract log messages into separate files
