+++ src/lib/utils/logger/messages/pieces-loader.messages.ts
@@ -1,0 +1,5 @@
+import { PieceName } from '../../pieces/pieces-loader';
+
+export function loadWarnMessage(pieceName: PieceName): string {
+  return `${pieceName} store is empty.`;
+}
